# public/openapi.yaml
openapi: 3.1.0
info:
  title: Open Hood public API
  version: 1.1.0
  summary: Consumer advocate — VIN, quotes, shops, recalls, integrations. Native fetch only.
  description: >
    Public routes that exist on this origin. POST /api/agent runs bay tools then writes a
    Beginner or Expert brief. GET /api/integrations/status is the honesty matrix: public
    probes, weekend keys configured:false when missing (probed:skip when on), Stripe test
    packet Checkout, live packet only with an existing Price ID, escrow refused.
    GET /agent/api renders this spec. GET /api/openapi is the live contract.
    Not a shop booking API. Not TecDoc. Not a Carfax feed.
servers:
  - url: /
    description: Same origin as the bay
...
paths:
...
  /api/openapi:
    get:
      tags: [Spec]
      operationId: getOpenApi
      summary: OpenAPI 3.1 document
      parameters:
        - name: format
          in: query
          schema:
            type: string
            enum: [yaml, json]
      responses:
        "200":
          description: OpenAPI 3.1
...
  /api/stripe/packet:
    get:
      tags: [Integrations]
      summary: Packet fee status
      description: Optional print packet / hold this bay. Test vs live vs none. Escrow refused.
      responses:
        "200":
          description: Status or session
    post:
      tags: [Integrations]
      summary: Open packet Checkout
      description: sk_test_ or sk_live_ plus existing price_. Founder can complete a test payment. Not escrow.
      responses:
        "200":
          description: Checkout URL
        "403":
          description: Live without product
        "404":
          description: No key
...